refactor: #4 라우터 정리
