+++ api/db/todolist.json
@@ -13,7 +13,7 @@
       "title": "TodoList 필터 추가",
       "content": "- sort\n- filter",
       "createdAt": "2023.11.16 14:40:07",
-      "updatedAt": "2023.11.16 16:54:03",
+      "updatedAt": "2023.11.17 13:35:58",
       "done": true
     }
   ],
