+++ react/src/Header.tsx
@@ -4,7 +4,7 @@ import React from "react";
 export default function Header() {
   return (
     <>
-    <h1>To Do List</h1>
+    <h1>할 일</h1>
     </>
   )
 }
